requirements-dev.txt
pytest==8.0.0
python-dotenv==1.0.1